requirements.txt
streamlit
gspread
oauth2client
google-generativeai>=0.8.0
streamlit-mic-recorder